domain.taskを__init__から移動
